requirements.txt
# ── Core ──
openai>=1.30.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ── Document Parsing ──
pypdf>=4.0.0
pdfplumber>=0.11.0
olefile>=0.47
pyhwp>=0.6.1
python-hwp>=0.1.0

# ── Embedding & Vector DB ──
chromadb>=0.5.0
faiss-cpu>=1.8.0

# ── HuggingFace (시나리오 A) ──
transformers>=4.45.0
sentence-transformers>=3.0.0
torch>=2.2.0

# ── Reranking ──
FlagEmbedding>=1.2.0

# ── Data Processing ──
pandas>=2.2.0
numpy>=1.26.0

# ── Evaluation ──
ragas>=0.2.0
rouge-score>=0.1.2
nltk>=3.9.0

# ── Streamlit App ──
streamlit>=1.30.0

# ── Notebook & Utils ──
jupyter>=1.0.0
ipywidgets>=8.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
